remove: delete utils file
